Fixes in pre-processing for issue #170
Do not force return type of StringBuilder functions

This was causing some type problems because the return types and the
lhs do not match.
As a consequence some equalities were ignored by the string solver.

Get rid of typecasts

These are not needed and may result in incorrect type later.